move exercise-related scripts to own folder
add exercise utils file
